requirements.txt
streamlit
pandas
yfinance
feedparser
requests
groq
python-dotenv
lxml